- fix compiler warnings

+++ util/data/dname.c
@@ -549,7 +549,7 @@ dname_lab_startswith(uint8_t* label, char* prefix, char** endptr)
 int
 dname_has_label(uint8_t* dname, size_t dnamelen, uint8_t* label)
 {
-	int len = *dname;
+	size_t len = *dname;
 	while(*dname && len <= dnamelen) {
 		if(*dname == *label && memlowercmp(dname, label, *dname) == 0)
 			return 1;
